Animator bugs fixed + Pet can sit on couch
Bug fixed: When in an animation and the "BallMode" button is pressed, the animation gets saved and the next time u press the animation button the animation continued playing, this is fixed now.

Pet can now sit on the couch by pressing on the couch.

Further improvement on resetting the pet's position when it has been rolled over and wants to move around